requirements.txt
scrapy
pandas
sqlalchemy
uuid
python-dotenv